Add plot utils for plotting
